read dbt_project.yml

+++ src/main.ts
@@ -1,8 +1,10 @@
 import * as fs from "fs/promises";
 import * as core from "@actions/core";
 import * as process from "process";
+import * as yaml from "js-yaml";
 import { exec, moveTo } from "./utils";
 import { Manifest } from "./manifest";
+import { isDBTProjectYml } from "./types";
 
 export async function main() {
   const moveBack = moveTo(core.getInput("dbt-project"));
@@ -26,26 +28,35 @@ export async function main() {
   core.setOutput("filepath", outpath);
 }
 
-const dummyProfile = {
-  dbt_mermaid: {
-    target: "dev",
-    outputs: {
-      dev: {
-        type: "postgres",
-        host: "postgres",
-        port: 5432,
-        dbname: "postgres",
-        schema: "main",
-        user: "postgres",
-        password: "password",
+function dummyProfile(profile: string) {
+  return {
+    dbt_mermaid: {
+      target: "dev",
+      outputs: {
+        dev: {
+          type: "postgres",
+          host: "postgres",
+          port: 5432,
+          dbname: "postgres",
+          schema: "main",
+          user: "postgres",
+          password: "password",
+        },
       },
     },
-  },
-};
+  };
+}
 
 async function preprocess() {
   const dbtVersion = core.getInput("dbt-version");
   const profiles = "profiles.yml";
+  const obj = await fs
+    .readFile("./dbt_project.yml")
+    .then((buf) => buf.toString())
+    .then((str) => yaml.load(str));
+  if (!isDBTProjectYml(obj)) {
+    throw "cannot read profile name from dbt_project.yml";
+  }
 
   let cleanup = async () => await fs.unlink(profiles);
   await fs
@@ -57,7 +68,7 @@ async function preprocess() {
     })
     .catch(() => {}); // NOP
 
-  await fs.writeFile(profiles, JSON.stringify(dummyProfile));
+  await fs.writeFile(profiles, JSON.stringify(dummyProfile(obj.profile)));
   await exec(`pipx run --spec dbt-postgres==${dbtVersion} dbt deps`);
   await exec(`pipx run --spec dbt-postgres==${dbtVersion} dbt ls`);
